changed UserInterface::frame to be an int instead of a long (it was giving compilers warnings,, and  2 billion frames is more than enough)

+++ include/UserInterface.hpp
@@ -54,7 +54,7 @@ private:
 
 	std::shared_ptr<Spritesheet> uiWireframe;
 
-	long frame;
+	int frame;
 
 	const Uint8 *keyDown;
